+++ src/l500/CMakeLists.txt
@@ -5,6 +5,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/l500-depth.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-private.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-color.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/l535-color.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-device.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-motion.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-factory.cpp"
@@ -18,6 +19,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/l500-depth.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-private.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-color.h"
+        "${CMAKE_CURRENT_LIST_DIR}/l535-color.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-device.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-motion.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-factory.h"
