+++ src/l500/l500-color.cpp
@@ -154,19 +154,6 @@ namespace librealsense
         return color_ep;
     }
 
-    l535_color::l535_color( std::shared_ptr< context > ctx,
-                            const platform::backend_device_group & group )
-        : device( ctx, group )
-        , l500_device( ctx, group )
-        , l500_color( ctx, group )
-    {
-        get_color_sensor()->register_processing_block(processing_block_factory::create_id_pbf(RS2_FORMAT_Y411, RS2_STREAM_COLOR));
-        get_color_sensor()->register_processing_block(
-            { { RS2_FORMAT_Y411 } },
-            { { RS2_FORMAT_RGB8, RS2_STREAM_COLOR } },
-            []() { return std::make_shared<Y411_converter>(RS2_FORMAT_RGB8); });
-    }
-
     l500_color::l500_color(std::shared_ptr<context> ctx, const platform::backend_device_group & group)
         :device(ctx, group),
         l500_device(ctx, group),
