+++ src/l500/l500-color.h
@@ -47,14 +47,6 @@ namespace librealsense
         ivcam2::intrinsic_rgb read_intrinsics_table() const;
         std::vector<uint8_t> get_raw_extrinsics_table() const;
     };
-
-    class l535_color :public l500_color
-    {
-    public:
-        l535_color(std::shared_ptr<context> ctx,
-            const platform::backend_device_group& group);
-    };
-
     class l500_color_sensor
         : public synthetic_sensor
         , public video_sensor_interface
