+++ src/proc/color-formats-converter.h
@@ -62,14 +62,14 @@ namespace librealsense
         void process_function(byte * const dest[], const byte * source, int width, int height, int actual_size, int input_size) override;
     };
 
-    class Y411_converter : public functional_processing_block
+    class y411_converter : public functional_processing_block
     {
     public:
-        Y411_converter(const rs2_format & target_format)
+        y411_converter(const rs2_format & target_format)
             : functional_processing_block("Y411 Transform", target_format) {};
 
     protected:
-        Y411_converter(const char * name, const rs2_format & target_format)
+        y411_converter(const char * name, const rs2_format & target_format)
             : functional_processing_block(
                 name, target_format, RS2_STREAM_COLOR, RS2_EXTENSION_VIDEO_FRAME)
         {
